venue-crawler/requirements.txt
# EventFoundry Venue Crawler - Python Dependencies
# Purpose: Comprehensive Kochi venue data extraction and processing

# Core Web Scraping
beautifulsoup4==4.12.3
lxml==5.1.0
html5lib==1.1

# Browser Automation (for dynamic content)
playwright==1.41.2
selenium==4.17.2

# HTTP Requests
requests==2.31.0
httpx==0.26.0
aiohttp==3.9.3

# Scraping Framework
scrapy==2.11.1

# Fuzzy Matching for Search Engine
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0

# Data Processing
pandas==2.2.0
openpyxl==3.1.2

# API Integration
google-api-python-client==2.116.0
googlemaps==4.10.0

# Rate Limiting & Proxy
ratelimit==2.2.1
requests-ratelimiter==0.4.2

# Data Validation
pydantic==2.6.1
jsonschema==4.21.1

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
tqdm==4.66.1

# Database (if needed for local storage)
pymongo==4.6.1
